+++ MGen/GenCF2.h
@@ -8,6 +8,6 @@ public:
 	CGenCF2();
 	~CGenCF2();
 	void Generate() override;
-	void LoadConfigLine(CString * sN, CString * sV, int idata, double fdata);
+	void LoadConfigLine(CString * sN, CString * sV, int idata, float fdata);
 };
